+++ lib/stoplight/notifier/base.rb
@@ -3,7 +3,10 @@
 module Stoplight
   module Notifier
     class Base
-      def notify(_message)
+      # @param _light [Light]
+      # @param _from_color [String]
+      # @param _to_color [String]
+      def notify(_light, _from_color, _to_color)
         fail NotImplementedError
       end
     end