+++ lib/stoplight/notifier/hip_chat.rb
@@ -4,7 +4,7 @@ module Stoplight
   module Notifier
     # @note hipchat ~> 1.3.0
     class HipChat < Base
-      DEFAULT_FORMAT = '@all %s'
+      DEFAULT_FORMAT = '@all Switching %s from %s to %s'
       DEFAULT_OPTIONS = { color: 'red', message_format: 'text', notify: true }
 
       # @param client [HipChat::Client]
@@ -17,8 +17,9 @@ module Stoplight
         @options = DEFAULT_OPTIONS.merge(options)
       end
 
-      def notify(message)
-        @client[@room].send('Stoplight', @format % message, @options)
+      def notify(light, from_color, to_color)
+        message = format(@format, light.name, from_color, to_color)
+        @client[@room].send('Stoplight', message, @options)
       end
     end
   end