+++ spec/stoplight/notifier/standard_error_spec.rb
@@ -10,20 +10,26 @@ describe Stoplight::Notifier::StandardError do
   after { $stderr = @stderr }
 
   describe '#notify' do
-    subject(:result) { notifier.notify(message) }
-    let(:message) { SecureRandom.hex }
+    subject(:result) { notifier.notify(light, from_color, to_color) }
+    let(:light) { Stoplight::Light.new(light_name, &light_code) }
+    let(:light_name) { SecureRandom.hex }
+    let(:light_code) { -> {} }
+    let(:from_color) { Stoplight::DataStore::COLOR_GREEN }
+    let(:to_color) { Stoplight::DataStore::COLOR_RED }
 
     it 'emits the message as a warning' do
       result
-      expect($stderr.string).to eql("#{message}\n")
+      expect($stderr.string)
+        .to eql("Switching #{light.name} from #{from_color} to #{to_color}\n")
     end
 
     context 'with a format' do
-      let(:format) { '> %s <' }
+      let(:format) { '%s %s %s' }
 
       it 'formats the message' do
         result
-        expect($stderr.string).to eql("> #{message} <\n")
+        expect($stderr.string)
+          .to eql("#{light.name} #{from_color} #{to_color}\n")
       end
     end
   end